ops-agent-core/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==4.0.1
cachetools==6.2.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
dataclasses-json==0.6.7
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.110.0
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
invoke==2.2.1
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.53
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
langsmith==0.1.147
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==1.26.4
openai==1.109.1
openpyxl==3.1.5
orjson==3.11.5
packaging==23.2
pandas==3.0.0
paramiko==4.0.0
passlib==1.7.4
pgvector==0.2.5
propcache==0.4.1
psycopg2-binary==2.9.9
pycparser==3.0
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.9
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
tenacity==8.5.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.27.1
yarl==1.22.0
zhipuai==2.1.5.20250825
captcha==0.7.1
python-jose[cryptography]==3.5.0
nacos-sdk-python==0.1.12
